requirements.txt
Flask
Spotipy
python-dotenv